refactor: change directory rendering components
