+++ examples/image/edge.html
@@ -6,7 +6,7 @@
     </body>
     <script src="../../dist/webgl-example.bundle.js"></script>
     <script>
-        const EdgeImageWindow = webglexample.rendering.core.window.EdgeImageWindow;
+        const EdgeImageWindow = webglexample.rendering.window.EdgeImageWindow;
         let element;
         let imageWindow;
         let image;
