+++ examples/image/magnifier.html
@@ -6,7 +6,7 @@
     </body>
     <script src="../../dist/webgl-example.bundle.js"></script>
     <script>
-        const MagnifierImageWindow = webglexample.rendering.core.window.MagnifierImageWindow;
+        const MagnifierImageWindow = webglexample.rendering.window.MagnifierImageWindow;
         let element;
         let imageWindow;
         let image;
