+++ examples/image/sharpen.html
@@ -6,7 +6,7 @@
     </body>
     <script src="../../dist/webgl-example.bundle.js"></script>
     <script>
-        const SharpenImageWindow = webglexample.rendering.core.window.SharpenImageWindow;
+        const SharpenImageWindow = webglexample.rendering.window.SharpenImageWindow;
         let element;
         let imageWindow;
         let image;
