+++ examples/image/smooth.html
@@ -6,7 +6,7 @@
     </body>
     <script src="../../dist/webgl-example.bundle.js"></script>
     <script>
-        const SmoothImageWindow = webglexample.rendering.core.window.SmoothImageWindow;
+        const SmoothImageWindow = webglexample.rendering.window.SmoothImageWindow;
         let element;
         let imageWindow;
         let image;
