+++ src/action/ImageEffectAction/index.ts
@@ -1,5 +1,5 @@
-import AbstractImageRenderer from '../../rendering/core/renderer/AbstractImageRenderer';
-import AbstractRenderer from '../../rendering/core/renderer/AbstractRenderer';
+import AbstractImageRenderer from '../../rendering/renderer/AbstractImageRenderer';
+import AbstractRenderer from '../../rendering/renderer/AbstractRenderer';
 import AbstractAnimationAction from '../AbstractAnimationAction';
 
 class ImageEffectAction extends AbstractAnimationAction {
