--- src/rendering/core/event/index.ts
@@ -1,5 +1,0 @@
-import EventCallback from './EventCallback';
-
-export default {
-  EventCallback,
-};
