+++ src/rendering/core/index.ts
@@ -1,13 +1,5 @@
-import camera from './camera';
-import event from './event';
-import renderer from './renderer';
 import Shader from './Shader';
-import window from './window';
 
 export default {
-  camera,
-  event,
-  renderer,
   Shader,
-  window,
 };
