+++ src/rendering/renderer/AbstractImageRenderer/index.ts
