+++ src/rendering/renderer/EdgeImageRenderer/glsl/fs.glsl
