+++ src/rendering/renderer/EdgeImageRenderer/glsl/vs.glsl
