+++ src/rendering/renderer/EdgeImageRenderer/index.ts
