+++ src/rendering/renderer/MagnifierImageRenderer/glsl/fs.glsl
