+++ src/rendering/renderer/MagnifierImageRenderer/glsl/vs.glsl
