+++ src/rendering/renderer/MagnifierImageRenderer/index.ts
