+++ src/rendering/renderer/SharpenImageRenderer/glsl/fs.glsl
