+++ src/rendering/renderer/SharpenImageRenderer/glsl/vs.glsl
