+++ src/rendering/renderer/SharpenImageRenderer/index.ts
