+++ src/rendering/renderer/SmoothImageRenderer/glsl/fs.glsl
