+++ src/rendering/renderer/SmoothImageRenderer/glsl/vs.glsl
