+++ src/rendering/renderer/SmoothImageRenderer/index.ts
