+++ src/rendering/window/EdgeImageWindow/index.ts
