+++ src/rendering/window/MagnifierImageWindow/index.ts
