+++ src/rendering/window/SharpenImageWindow/index.ts
