+++ src/rendering/window/SmoothImageWindow/index.ts
