+++ src/rendering/window/index.ts
